refactor: extract resize observation logic into useResizeObserver hook
- Create a reusable `useResizeObserver` hook in `src/hooks` to observe element width.
- Encapsulate debouncing and immediate update logic for initial visibility within the hook.
- Refactor `BookGrid.tsx` to use the new hook, reducing component complexity and line count.
- Improve type safety by specifying `HTMLDivElement` for the container reference.
